Move GenerateCommandConfigurationFactory
